feat(cli): add version command to display application version
This commit introduces the `version` command to the CLI interface, allowing users to display the current version of the application. A new rule has been added to the `cli.ts` file to recognize the `version` command along with its description. Additionally, a new subcommand file `version.ts` has been created which implements the functionality to output the version details. This command supports an optional `--json` flag to output the version information in JSON format, which can be useful for integration with other tools or scripts. This enhancement improves the usability of the CLI by providing users with an easy way to check the version of the application they are working with.