docs(license): add MIT License and update package.json
This commit introduces the MIT License to the project, establishing clear permissions for software redistribution, modification, and use in both private and commercial contexts. By selecting the MIT License, we embrace an open-source philosophy, encouraging greater collaboration and usage flexibility. Additionally, the `package.json` file has been updated to reflect this licensing choice. This amendment aligns with our commitment to transparency and supports the community's ability to engage with and contribute to the project under defined, permissive terms. Including the license directly within the repository ensures that all copies or substantial portions of the software will carry the same license, simplifying legal requirements for derivatives and distributions.