docs(README): document .env source customization and best practices
This commit enhances the README.md with a new section titled "Sources," which elaborates on the variety of source file formats and locations that can be utilized to populate the `.env` file. It explicitly lists supported filenames and paths, such as `.env.foo`, `.foo.env`, and variations within an `.envs/` directory. This addition aims to provide users with flexible options for managing environment variables, catering to different organizational or project-specific standards.

To guide best practices, the documentation now advises users to add the `.envs/` directory to their `.gitignore` file. This recommendation is crucial for security, as it helps prevent the accidental publication of sensitive information contained within environment variable files.

By clarifying these options and encouraging secure practices, this commit aims to improve usability and security for users leveraging the described tool for environment variable management.